Move getting attachment to EmailRestController
